requirements.txt
pynmodl
textx
brute-curvefit
numpy
neuron
matplotlib
pymoose